fix XcodeList for worspaces and redesing it's logic
move XcodeList creation to XcodeBuild type
handle 'xcodebuild -list -json' for workspaces
right now only handle them, building for workspaces isn't implemented
